added Validation for Shows , query parameter for read_shows

+++ Shows/controller.py
@@ -1,45 +1,24 @@
 from flask import jsonify , request
 from Utilities.db_operations import *
+from Utilities.error_handler import custom_response_maker
+from .validation import *
 
-def custom_response_maker(res):
-    message = ''
-    is_success = False
-
-    if 'err_code' in res:
-        print("response in crm: ",res)
-        return jsonify({
-            "is_success" : is_success,
-            "data" : None,
-            "message": res['message'],
-            "start_time": None,
-            "end_time": None,
-            "duration": None
-        })
-    elif isinstance(res["data"],int):
-        message = "no of rows changed : %d"%(res["data"])
-        is_success = True
-    elif isinstance(res["data"],list):
-        is_success = True
-
-    return jsonify({
-        "is_success" : is_success,
-        "data" : res["data"],
-        "message": message,
-        "start_time": res["start_time"],
-        "end_time": res["end_time"],
-        "duration": res["end_time"] - res["start_time"]
-    })
-    
 def home():
     return "<h1>This is the First Page</p1>"
 
 def read_from_db():
-    query = 'SELECT * FROM TvShows'
+    if len(request.args)>0:
+        params = request.args
+        query = 'SELECT * FROM TvShows WHERE %s="%s"'%(params['column'],params['value'])
+    else:
+        query = 'SELECT * FROM TvShows'
     return custom_response_maker(read_row(query))
 
 def insert_to_db():
     body = {}
     body = request.get_json()
+    if (isinstance(validate_insert(body), dict)):
+        return custom_response_maker(validate_insert(body))
     try:
         query = 'INSERT INTO TvShows VALUES (%d , "%s" , "%s" );'%(
                 body['id'], body['name'], body['studio'])
@@ -54,6 +33,8 @@ def insert_to_db():
 def update_to_db():
     body = {}
     body = request.get_json()
+    if (isinstance(validate_update(body), dict)):
+        return custom_response_maker(validate_insert(body))
     try:
         query = 'UPDATE TvShows SET %s="%s" WHERE id = %d;'%(
                 body['column'], body['new_value'], body['id'])
@@ -68,6 +49,8 @@ def update_to_db():
 def delete_from_db():
     body = {}
     body = request.get_json()
+    if (isinstance(validate_delete(body), dict)):
+        return custom_response_maker(validate_insert(body))
     try:
         query = 'DELETE FROM TvShows WHERE id = %d ;'%(body['id'])
     except KeyError as e:
